requirements.txt
langchain
python-dotenv
ipykernel
langchain_community
pypdf
bs4
arxiv
pymupdf
wikipedia
langchain-text-splitters
langchain-openai
langchain-huggingface
sentence_transformers
faiss-cpu
chromadb
langchain-chroma
streamlit
langchain-groq
langserve
fastapi
uvicorn
duckduckgo-search